Refactor schema builders into own modules
