--- aquadoggo/src/dynamic_graphql/queries/mod.rs
@@ -1,5 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-
-mod next_args;
-
-pub use next_args::next_args;
