+++ aquadoggo/src/dynamic_graphql/schema_builders/mod.rs
@@ -1,0 +1,9 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+mod document_field;
+mod next_args;
+mod document;
+
+pub use next_args::build_next_args_query;
+pub use document_field::build_document_field_schema;
+pub use document::{build_document_schema, build_document_query};
